WX-1306 workflow submission working against static workspace, need to get workflow fetch running
