+++ server/src/test/python/cromwell-az-e2e-test/cromwell_az_e2e_test/workflow_files/hello.inputs
