+++ server/src/test/python/cromwell-az-e2e-test/pyproject.toml
@@ -1,5 +1,5 @@
 [tool.poetry]
-name = "cromwell-az-e2e"
+name = "cromwell-az-e2e-test"
 version = "0.1.0"
 description = ""
 authors = ["JVThomas <jthomas0209@gmail.com>"]
@@ -15,3 +15,6 @@ line-length = 120
 [build-system]
 requires = ["poetry-core"]
 build-backend = "poetry.core.masonry.api"
+
+[tool.poetry.scripts]
+start = "cromwell_az_e2e_test.az_e2e:start"
